Subgrid support
Refactoring to allow a single puzzle entry to have several subgrids.

This allows a puzzle like an acrostic to have multiple lines,
but only one set of commands and a shared undo.

Adds infrastructure for separate grids to be linked together,
e.g. if a puzzle consists of draggable parts.
Note that the infrastructure alone is not enough to expose this, but
it's a good base for future experimentation.

Also updated copyright dates.