Validator structure tweak
Tweaking the validator structure to allow validators to be written with less code and also describe themselves (future UX will use these descriptions).

Validators are no longer functions; they are simple Javascript objects that expose two functions, one to get the description and one to validate.

The validation function no longer returns a result of -1/0/1; simply calling incomplete() or fail() on any objects exposed from puzzleGrid will automatically mark the puzzle as incomplete or failed.

Also renaming the edge/path/spoke concept of "segment" to "chain", as "segment" was kind of a misnomer. A "chain" is a simple series of segments with two endpoints, as opposed to a "loop" which has no endpoints and a "web" which has a complex structure.